requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
rapidfuzz
pandas
openpyxl
psycopg[binary]
alembic
passlib[bcrypt]==1.7.4
python-dotenv
bcrypt==3.2.2S